requirements.txt
asgiref==3.8.1
Django==5.1.5
joblib==1.4.2
liac-arff==2.5.0
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0
pytz==2024.1
scikit-learn==1.6.1
scipy==1.13.1
sqlparse==0.5.0
tzdata==2024.1
gunicorn==23.0.0
django-cors-headers==4.6.0
whitenoise==6.8.2